Moving toward more FloatingPoint conformance.
Also removed commented out unit test that was testing if a Double obtained from a WideFloat created from a UInt64 was as close or closer to the original UInt64 than a Double created directly from the UInt64.  That test was wrong-headed.  After a lot of debugging,  we store more significant digits, which do rounding.  Then when we obtain a Double from the WideFloat, there's more rounding.  This results, sometimes results in a Double that is further away from the the original UInt64 than a Double created directly from it, which does only one rounding.   If instead, you obtain a UInt64 from a WideUInt created from a UInt64, and get a UInt64 from a Double created with the same UInt64, our value is indeed closer to the original.   So the problem with the test was that it was comparing a value resulting from rounding twice to a value that was rounded only once.  The circumstances when that fails are rare (requiring all 1s in the least significant 10 bits), but they can happen and don't indicate anything amiss with WideFloat.